requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
arabic-reshaper==3.0.0
argcomplete==3.6.2
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
datasets==4.2.0
decorator==5.2.1
Deprecated==1.3.1
dill==0.4.0
distro==1.9.0
docopt==0.6.2
exceptiongroup==1.3.0
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==4.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.59.2
fpdf==1.7.2
frozenlist==1.8.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
Js2Py==0.74
kiwisolver==1.4.9
limits==5.8.0
llvmlite==0.39.1
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.6
more-itertools==10.8.0
moviepy==2.2.1
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
networkx==3.4.2
noisereduce==3.0.3
numba==0.56.4
numpy==2.2.6
openai==1.108.1
openai-whisper==20250625
opencv-contrib-python==4.12.0.88
ordered-set==4.1.0
packaging==25.0
pandas==2.3.2
pedalboard==0.9.22
pillow==11.3.0
pipwin==0.5.2
playsound==1.2.2
proglog==0.1.12
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.1
psycopg2-binary==2.9.10
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pybind11==3.0.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
pyjsparser==2.7.1
PyJWT==2.10.1
pyparsing==3.2.3
PyPrind==2.11.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
pySmartDL==1.3.4
python-barcode==0.16.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-escpos==3.1
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzbar==0.1.9
QDarkStyle==3.2.3
qrcode==8.2
QtPy==2.4.3
redis==7.4.0
regex==2025.9.1
reportlab==4.4.3
requests==2.32.5
rsa==4.9.1
sacremoses==0.1.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.8
SpeechRecognition==3.14.3
SQLAlchemy==2.0.43
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
tomli==2.2.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==2.1.2
xxhash==3.6.0
yarl==1.22.0
yt-dlp==2026.3.17
